requirements.txt
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
DateTime==4.2
dj-database-url==0.5.0
Django==2.0.6
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.8.1
httpie==0.9.9
idna==2.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
Pygments==2.2.0
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
whitenoise==3.3.1
zope.interface==4.5.0
